refactor: having the child manage movies didn't make sense, so the handler got defined in the parent and passed to the child only.
